添加一些打印，此版本有一个bug
bug:在拖入一个元件，进行移动，回退，再撤销，关闭会出现移除，感觉是两次delete

+++ src/APP/AppMainWindow.cpp
@@ -135,8 +135,10 @@ void AppMainWindow::closeEvent(QCloseEvent* e)
         }
     } else {
         // 不保存要删除
-        if (!QFile::remove(uistateFile)) {
-            qDebug() << tr("can not remove %1").arg(uistateFile);
+        if (QFile::exists(uistateFile)) {
+            if (!QFile::remove(uistateFile)) {
+                qDebug() << tr("can not remove %1").arg(uistateFile);
+            }
         }
     }
     // Delete dock manager here to delete all floating widgets. This ensures
